+++ app/(auth)/login/page.tsx
@@ -8,7 +8,7 @@ import { toast } from "sonner";
 import Link from "next/link";
 import axios from "axios";
 import TurnstileComponent from "@/components/ui/turnstile";
-import verifyTurnstileToken from "@/lib/helpers";
+import verifyTurnstileToken from "@/lib/actions/verifycaptcha";
 type Props = {};
 
 interface FormData {
@@ -194,7 +194,7 @@ export default function Login({}: Props) {
             <button
               className="inline-flex items-center justify-center text-xl font-medium border disabled:pointer-events-none disabled:bg-green-50 disabled:text-black  h-10 px-4 py-2 w-full bg-green-500 hover:bg-green-600 text-white rounded-md"
               type="submit"
-              disabled={loading}
+              disabled={loading || !token}
               title="login">
               {!loading ? (
                 "Login"
